+++ R/flowGate-package.R
@@ -1,0 +1,9 @@
+#' @description
+#' To learn more about how to use flowGate, please start by browsing the vignettes:
+#'
+#' @keywords internal
+"_PACKAGE"
+
+## usethis namespace: start
+## usethis namespace: end
+NULL