+++ NEWS.md
@@ -1,2 +1,9 @@
 # Changes in version 0.99.3 (2023-03-15)
  - Document cleanup as part of bioconductor review
+
+# Changes in version 1.13.1 (2026-07-02)
+ - Implemented fixes due to the ggplot2 version 4 changes, with 
+ the update ggcyto syntax changes
+ - Where appropiate, switched from import to importFrom tags
+ - Implemented gs_apply_gate_check to allow adjustment existing gates
+ - Additional unit tests to test suite